Responsive updates and storybook clean up
