+++ draft-packages/modal/KaizenDraft/Modal/Primitives/GenericModalSection.scss
@@ -1,11 +1,6 @@
 @import "~@kaizen/design-tokens/sass/color";
 @import "~@kaizen/design-tokens/sass/border";
-@import "~@kaizen/design-tokens/sass/spacing";
-@import "~@kaizen/deprecated-component-library-helpers/styles/type";
 
-// .padded {
-//   padding: $spacing-lg;
-// }
 .inputEdit {
   background-color: $color-gray-200;
   border-radius: $border-solid-border-radius;
