+++ draft-packages/modal/KaizenDraft/Modal/Primitives/ModalAccessibleDescription.scss
@@ -1,3 +1,4 @@
 .modalDescription {
   grid-column-start: 2;
+  padding-bottom: 1.5rem;
 }
