+++ draft-packages/modal/KaizenDraft/Modal/Primitives/ModalFooter.scss
@@ -1,6 +1,7 @@
 @import "~@kaizen/deprecated-component-library-helpers/styles/layout";
 @import "~@kaizen/deprecated-component-library-helpers/styles/type";
 @import "~@kaizen/deprecated-component-library-helpers/styles/color";
+@import "~@kaizen/component-library/styles/responsive";
 
 .actions {
   display: flex;
@@ -15,13 +16,26 @@
 
 .padded {
   padding: 0 $ca-grid * 1.5 $ca-grid * 1.5 $ca-grid * 1.5;
-  &.informationPadded {
-    padding: 1.25rem $ca-grid * 1.5 0 $ca-grid * 1.5;
+  @include ca-media-mobile {
+    padding: $ca-grid;
+    flex-direction: column;
   }
 }
-
+.actionButton {
+  @include ca-media-mobile {
+    width: 100%;
+    > *,
+    > * > * {
+      width: 100%;
+      justify-content: center;
+    }
+  }
+}
 .actionButton + .actionButton {
   @include ca-margin($end: $ca-grid * 0.5);
+  @include ca-media-mobile {
+    @include ca-margin($top: $ca-grid * 0.5);
+  }
 }
 
 // to avoid the row-reverse styles
