+++ packages/design-tokens/docs/StoryBoard.stories.tsx
@@ -35,7 +35,9 @@ import * as IllustrationSpotStories from "@kaizen/draft-illustration/docs/Illust
 import * as LoadingPlaceholderStories from "@kaizen/draft-loading-placeholder/docs/LoadingPlaceholder.stories"
 import * as LoadingSpinnerStories from "@kaizen/draft-loading-spinner/docs/LoadingSpinner.stories"
 import * as MenuStories from "@kaizen/draft-menu/docs/Menu.stories"
-import * as ModalStories from "@kaizen/draft-modal/docs/Modal.stories"
+import * as InputEditModalStories from "@kaizen/draft-modal/docs/InputEditModal.stories"
+import * as ConfirmationModalStories from "@kaizen/draft-modal/docs/ConfirmationModal.stories"
+import * as ContextModalStories from "@kaizen/draft-modal/docs/ContextModal.stories"
 import * as PageLayoutStories from "@kaizen/draft-page-layout/docs/PageLayout.stories"
 import * as RadioGroupStories from "@kaizen/draft-form/docs/RadioGroup.stories"
 import * as PopoverStories from "@kaizen/draft-popover/docs/Popover.stories"
@@ -442,7 +444,9 @@ export const Everything: Story = () => {
         <StoriesContainer storyModule={LoadingPlaceholderStories} />
         <StoriesContainer storyModule={LoadingSpinnerStories} />
         <StoriesContainer storyModule={MenuStories} />
-        <StoriesContainer storyModule={ModalStories} />
+        <StoriesContainer storyModule={InputEditModalStories} />
+        <StoriesContainer storyModule={ConfirmationModalStories} />
+        <StoriesContainer storyModule={ContextModalStories} />
         <StoriesContainer storyModule={PageLayoutStories} />
         <StoriesContainer storyModule={ParagraphStories} />
         <StoriesContainer storyModule={PopoverStories} />
